SRB2: merged ALAM's hell
git-svn-id: svn://code.srb2.org/SRB2/trunk@4572 d18046d3-e9af-4449-a702-ac2b6c676ba2

+++ src/win32/win_vid.c
@@ -445,8 +445,8 @@ void I_FinishUpdate(void)
 			/// optimized for p2 or mmx??
 			if (ScreenHeight > vid.height)
 			{
-				BYTE *ptr = ScreenPtr;
-				long half_excess = ScreenPitch*(ScreenHeight-vid.height)/2;
+				byte *ptr = ScreenPtr;
+				size_t half_excess = ScreenPitch*(ScreenHeight-vid.height)/2;
 				memset(ptr, 0x1F, half_excess);
 				ptr += half_excess;
 				VID_BlitLinearScreen(screens[0], ptr, vid.width*vid.bpp, vid.height,
